refactor : 패키지 분리 및 StationName 분리
